% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooper.R
\docType{package}
\name{cooper}
\alias{cooper}
\title{Package for MATH4753, OU, Fall 2020}
\description{
A collection of functions pulled from various lab assignments for the MATH 4753 class at OU for the 2020 Fall Semester.
}
\author{
Zachary Cooper \email{zcoop@ou.edu}
}
