% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Generate Samples from a given Sample, "bootstrap-wise," and create an accompanying detailed histograph}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Numeric, x Numeric, fun String, alpha Numeric, cx Numeric.}
}
\value{
List of the calculated confidence interval, the function corresponding to the calculated statistic, and the input sample.
}
\description{
You have at least 5 inputs.
'iter' is the number of samples you want generated
'x' is the input sample. Should be a vector of numbers.
'fun' is the function that corresponds with the statistic you want each 'iter' sample to calculate. Some common inputs would be "mean" "sd" "median".)
'alpha' is 1 minus the confidence interval you want calculated. So if you want a 95% Confidence Interval, 'alpha = 0.05'.
'cx' is merely a variable that should be a numerical input that is used with the 'cex' parameter in making the histograph. It determines how large the symbols and letter in the graph are.
The output is a histograph of the statistic of each of the 'iter' samples.
Additionally, a vertical line showing the value of that statistic for the input sample, and labels displaying the confidence interval.
Additionally, 'myboot2()' will return a list containing the calculated confidence interval, the function corresponding to the calculated statistic, and the input sample. Note, this list is 'invisible' in that it won't display by default but only if explicitly invoked (to cut back on potentially displaying large screeds of data).
}
\examples{
myboot2(10000, rnorm(20, mean = 10, sd = 4), fun = "mean", alpha = 0.05, cx = 1.5)

}
