% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{95% Confidence Interval for mean of 1 normally distributed, random sample}
\usage{
myci(x)
}
\arguments{
\item{x}{Numeric.}
}
\value{
Array(lower bound of ci, upper bound of ci)
}
\description{
You have 1 input.
'x' is your data. It should be an array of numbers corresponding to some normally distributed, random sample.
The output is an array of 2 values. The first value is the lower bound of the 95% confidence interval.
And the second value is the upper bound of the 95% confidence interval.
}
\examples{
myci(rnorm(25, mean = 10, sd = 5))

}
