% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Creates Samples from a Uniform Distribution, Calc's the Mean Statistic w/ Histograph}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{Numeric, iter Numeric, a Numeric, b Numeric.}
}
\value{
Histograph of the means of each sample (with 3 additional curves overlaid)
}
\description{
You have 4 inputs.
'n' is the size of each sample
'iter' is the number of samples you want generated
'a' is the lowerbound of the uniform distribution (that the samples will be pulled from)
'b' is the upperbound of the uniform distribution (that the samples will be pulled from)
The output is a histograph of the means of each sample (so there are 'iter' many datapoints for the graph).
Additionally, 3 curves are overlaid on the graph:
- a density curve made from the sample distribution
- a theoretical normal curve
- the density from which the samples were taken
}
\examples{
mycltu(30, 10000, 0, 10)

}
