% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Creates a piecewise function of two lines connected at x=18 in order to make a continuous curve.}
\usage{
myf(x, coef)
}
\arguments{
\item{x}{Numeric, coef Numeric Vector (of at least 3 elements).}
}
\value{
Numeric.
}
\description{
You feed the function an x-value and a vector of 3 numbers to be the 3 coefficients, and it'll calculate a value and return that value.
If x less than or equal to 18, then the function will return coef[1]+coef[2]*x.
If x > 18, then the function will return coef[1]+coef[2]*x+coef[3]*(x-18)
}
\examples{
beta = c(1,2,3)
myf(0, beta)

}
