% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{Maximum Likelihood Solution for a distribution via log-likelihood}
\usage{
mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{String, theta Numeric.}
}
\value{
Numerical value of theta where the maximum point of the distribution is located.
}
\description{
You have 2 inputs.
'lfun' is the log likelihood equation that the function will use to solve for its max value.
'theta' is the function parameter which will typically be an array in the form of 'seq(a, b, length = c)' where you guess the interval (a, b) that contains the maximum value, and 'c' is how fine of a resolution it will step through in checking for maximum values.
The output is a plot of the original function (before taking the log of it) with a vertical line overlaid where it finds the maximum value of this curve.
Additionally, the input (which is a value of theta) that results in the maximum value is returned.
}
\examples{
mymaxlikg(lfun = "logbin2", theta = seq(0, 1, length = 10000))

}
