% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbin.R
\name{mynbin}
\alias{mynbin}
\title{Negative Binomial Probability at a Point}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{y}{Numeric, r Numeric, p Numeric.}
}
\value{
Numeric.
}
\description{
You have 3 inputs.
'p' is the probability that a single Bernoulli trial is successful.
'r' is how many successful trials you are looking for.
'y' is the number trials.
The output is the probability of 'r' number of successes occurring given 'y' number of trials.
}
\examples{
mynbin(y = 6, r = 4, p = 0.8)

}
