% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal distribution curve with lower-tail shading and calculation}
\usage{
myncurve(q, mu, sigma)
}
\arguments{
\item{q}{Numeric, mu Numeric, sigma Numeric.}
}
\value{
Graph of norm(mean = mu, sd = sigma), a red shaded polygon (representing a lower-tail probability) overlaid on the graph, and a list returning the qnorm(q, mu, sigma) value.
}
\description{
You have 3 inputs.
'q' is the quantile you're interested in finding the probability from (-infinity, q). This is also the horizontal location on the curve that the shaded area under the curve from x=-infinity to x=q.
'mu' is the mean of the curve (the horizontal location for the mid point of the symetric curve).
'sigma' is the standard deviation of the curve (how narrow or wide the curve behaves).
The output is the lower-tail probability that a random specimen's result will occur at less than or equal to 'q'.
}
\examples{
mynbin(10, 10, 4)

}
